NEI recipe import (the '?' button) support
Lack of this feature was bugging the hell out of me so I decided to add
it.

Normal left click functionality, showing a shadow recipe on the crafting
grid, works flawlessly.

Shift+left click function, that normally figures the max amount of the
item you can make with your inventory items and pulls all of those item
stacks to the crafting grid, is a bit of an issue. Now it pulls items
only from player inventory and not internal inventory, would need to
implement new overlay handler class to make it do that.